geminiApi/asset/pyproject.toml
[project]
name = "asset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopy>=2.4.1",
    "google-cloud-aiplatform>=1.72.0",
]

[tool.setuptools]
py-modules = ["index", "service", "template"]
